# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MindGuard API",
    "description": "Autonomous multi-agent AI attention firewall. All actions are proposals until authorized by the deterministic guardrail engine.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/interventions/evaluate": {
      "post": {
        "tags": [
          "Interventions",
          "Feedback",
          "Risk",
          "Content"
        ],
        "summary": "Evaluate",
        "operationId": "evaluate_api_interventions_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/interventions/feedback": {
      "post": {
        "tags": [
          "Interventions",
          "Feedback",
          "Risk",
          "Content"
        ],
        "summary": "Feedback",
        "operationId": "feedback_api_interventions_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/interventions": {
      "get": {
        "tags": [
          "Interventions",
          "Feedback",
          "Risk",
          "Content"
        ],
        "summary": "List Interventions",
        "operationId": "list_interventions_api_interventions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Interventions Api Interventions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interventions/{intervention_id}": {
      "get": {
        "tags": [
          "Interventions",
          "Feedback",
          "Risk",
          "Content"
        ],
        "summary": "Read Intervention",
        "operationId": "read_intervention_api_interventions__intervention_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "intervention_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Intervention Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Read Intervention Api Interventions  Intervention Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/risk/history": {
      "get": {
        "tags": [
          "Interventions",
          "Feedback",
          "Risk",
          "Content"
        ],
        "summary": "Risk History",
        "operationId": "risk_history_api_risk_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Risk History Api Risk History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/risk/assess": {
      "post": {
        "tags": [
          "Interventions",
          "Feedback",
          "Risk",
          "Content"
        ],
        "summary": "Risk Assess",
        "description": "Dry run: context -> behaviour -> policy -> risk. Nothing is persisted and no intervention is created.",
        "operationId": "risk_assess_api_risk_assess_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskAssessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Risk Assess Api Risk Assess Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/content/analyze": {
      "post": {
        "tags": [
          "Interventions",
          "Feedback",
          "Risk",
          "Content"
        ],
        "summary": "Analyze Content",
        "operationId": "analyze_content_api_content_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentAssessment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/content/screenshot": {
      "post": {
        "tags": [
          "Interventions",
          "Feedback",
          "Risk",
          "Content"
        ],
        "summary": "Analyze Image",
        "operationId": "analyze_image_api_content_screenshot_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_image_api_content_screenshot_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentAssessment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_analyze_image_api_content_screenshot_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "app_package": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Package"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_image_api_content_screenshot_post"
      },
...
      "ContentAssessment": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ContentCategory",
            "default": "unknown"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 0.0
          },
          "goal_relevance": {
            "$ref": "#/components/schemas/GoalRelevance",
            "default": "neutral"
          },
          "source": {
            "$ref": "#/components/schemas/ContentSource",
            "default": "metadata"
          },
          "classifier": {
            "type": "string",
            "title": "Classifier",
            "default": "metadata"
          },
          "injection_detected": {
            "type": "boolean",
            "title": "Injection Detected",
            "default": false
          },
          "injection_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Injection Score",
            "default": 0.0
          },
          "injection_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Injection Signals"
          },
          "used_llm": {
            "type": "boolean",
            "title": "Used Llm",
            "default": false
          },
          "text_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Sha256"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ContentAssessment"
      },
...
      "ContentInput": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 40000,
            "minLength": 1,
            "title": "Text"
          },
          "source": {
            "$ref": "#/components/schemas/ContentSource",
            "default": "text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ContentInput"
      },
...
      "DecisionSource": {
        "type": "string",
        "enum": [
          "deterministic",
          "bandit",
          "llm",
          "fallback",
          "override"
        ],
        "title": "DecisionSource"
      },
      "DeviceCapabilities": {
        "properties": {
          "usage_access": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage Access"
          },
          "notifications": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notifications"
          },
          "overlay": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlay"
          },
          "dnd_access": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dnd Access"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DeviceCapabilities",
        "description": "What the user's device has granted. `None` = not reported (treated as available)."
      },
...
      "EvaluateRequest": {
        "properties": {
          "app_package": {
            "type": "string",
            "maxLength": 255,
            "pattern": "^[a-zA-Z][a-zA-Z0-9_]*(\\.[a-zA-Z0-9_]+)+$",
            "title": "App Package"
          },
          "session_minutes": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1440.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Minutes"
          },
          "content": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "device_capabilities": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceCapabilities"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "app_package"
        ],
        "title": "EvaluateRequest"
      },
      "EvaluationResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "decision": {
            "$ref": "#/components/schemas/InterventionType"
          },
          "proposed": {
            "$ref": "#/components/schemas/InterventionType"
          },
          "duration_minutes": {
            "type": "integer",
            "title": "Duration Minutes"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "decided_by": {
            "$ref": "#/components/schemas/DecisionSource"
          },
          "reason_codes": {
            "items": {
              "$ref": "#/components/schemas/ReasonCode"
            },
            "type": "array",
            "title": "Reason Codes"
          },
          "guardrail_flags": {
            "items": {
              "$ref": "#/components/schemas/ReasonCode"
            },
            "type": "array",
            "title": "Guardrail Flags"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "explanation_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Explanation Points"
          },
          "command": {
            "$ref": "#/components/schemas/DeviceCommand"
          },
          "intervention_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intervention Id"
          },
          "risk": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PolicySummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "content": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentAssessment"
              },
              {
                "type": "null"
              }
            ]
          },
          "degraded": {
            "type": "boolean",
            "title": "Degraded",
            "default": false
          },
          "path": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Path"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "decision",
          "proposed",
          "duration_minutes",
          "confidence",
          "decided_by",
          "reason_codes",
          "guardrail_flags",
          "explanation",
          "explanation_points",
          "command"
        ],
        "title": "EvaluationResponse"
      },
...
      "FeedbackRequest": {
        "properties": {
          "intervention_id": {
            "type": "string",
            "format": "uuid",
            "title": "Intervention Id"
          },
          "outcome": {
            "$ref": "#/components/schemas/OutcomeType"
          },
          "satisfaction": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Satisfaction"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "intervention_id",
          "outcome"
        ],
        "title": "FeedbackRequest"
      },
      "FeedbackResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "intervention_id": {
            "type": "string",
            "format": "uuid",
            "title": "Intervention Id"
          },
          "outcome": {
            "$ref": "#/components/schemas/OutcomeType"
          },
          "reward": {
            "type": "number",
            "title": "Reward"
          },
          "decision": {
            "$ref": "#/components/schemas/InterventionType"
          },
          "bandit_updated": {
            "type": "boolean",
            "title": "Bandit Updated"
          },
          "memory_written": {
            "type": "boolean",
            "title": "Memory Written"
          },
          "preference": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preference"
          },
          "suggestion": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestion"
          },
          "insights": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Insights"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "intervention_id",
          "outcome",
          "reward",
          "decision",
          "bandit_updated",
          "memory_written"
        ],
        "title": "FeedbackResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterventionType": {
        "type": "string",
        "enum": [
          "ALLOW",
          "SOFT_WARNING",
          "MINDFUL_PROMPT",
          "DELAY",
          "LIMITED_ACCESS",
          "TEMPORARY_BLOCK",
          "FOCUS_MODE",
          "REQUEST_CONFIRMATION"
        ],
        "title": "InterventionType"
      },
...
      "PolicySummary": {
        "properties": {
          "effect": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PolicyEffect"
              },
              {
                "type": "null"
              }
            ]
          },
          "winning_rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winning Rule Id"
          },
          "rule_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Description"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation",
            "default": ""
          },
          "in_restricted_window": {
            "type": "boolean",
            "title": "In Restricted Window",
            "default": false
          }
        },
        "type": "object",
        "title": "PolicySummary"
      },
...
      "RiskAssessRequest": {
        "properties": {
          "app_package": {
            "type": "string",
            "maxLength": 255,
            "pattern": "^[a-zA-Z][a-zA-Z0-9_]*(\\.[a-zA-Z0-9_]+)+$",
            "title": "App Package"
          },
          "session_minutes": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1440.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Minutes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "app_package"
        ],
        "title": "RiskAssessRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}